+++ Sources/Twift+Search.swift
@@ -31,8 +31,8 @@ extension Twift {
     if let nextToken = nextToken { queryItems.append(URLQueryItem(name: "next_token", value: nextToken)) }
     if let sinceId = sinceId { queryItems.append(URLQueryItem(name: "since_id", value: sinceId)) }
     if let untilId = untilId { queryItems.append(URLQueryItem(name: "until_id", value: untilId)) }
-    if let startTime = startTime?._ISO8601Format() { queryItems.append(URLQueryItem(name: "start_time", value: startTime)) }
-    if let endTime = endTime?._ISO8601Format() { queryItems.append(URLQueryItem(name: "end_time", value: endTime)) }
+    if let startTime = startTime?.ISO8601Format() { queryItems.append(URLQueryItem(name: "start_time", value: startTime)) }
+    if let endTime = endTime?.ISO8601Format() { queryItems.append(URLQueryItem(name: "end_time", value: endTime)) }
     
     let fieldsAndExpansions = fieldsAndExpansions(for: Tweet.self, fields: fields, expansions: expansions)
     
@@ -74,8 +74,8 @@ extension Twift {
     if let nextToken = nextToken { queryItems.append(URLQueryItem(name: "next_token", value: nextToken)) }
     if let sinceId = sinceId { queryItems.append(URLQueryItem(name: "since_id", value: sinceId)) }
     if let untilId = untilId { queryItems.append(URLQueryItem(name: "until_id", value: untilId)) }
-    if let startTime = startTime?._ISO8601Format() { queryItems.append(URLQueryItem(name: "start_time", value: startTime)) }
-    if let endTime = endTime?._ISO8601Format() { queryItems.append(URLQueryItem(name: "end_time", value: endTime)) }
+    if let startTime = startTime?.ISO8601Format() { queryItems.append(URLQueryItem(name: "start_time", value: startTime)) }
+    if let endTime = endTime?.ISO8601Format() { queryItems.append(URLQueryItem(name: "end_time", value: endTime)) }
     
     let fieldsAndExpansions = fieldsAndExpansions(for: Tweet.self, fields: fields, expansions: expansions)
